feat: simplify usage of twoslash inline cache
