+++ packages/vitepress-twoslash/src/cache-inline/index.ts
@@ -1,67 +1,39 @@
-import type { TwoslashTypesCache } from '@shikijs/twoslash'
-import type { ShikiTransformer } from 'shiki'
 import type { Plugin } from 'vite'
 import type { UserConfig } from 'vitepress'
-import { readFileSync } from 'node:fs'
+import type { VitePressPluginTwoslashOptions } from '../types'
+import { transformerTwoslash } from '..'
 import { createInlineTypesCache } from './cache-inline'
-import { createFenceSourceMap } from './fence-source-map'
+import { withFenceSourceMap } from './fence-source-map'
 
-export function createTwoslashInlineCache(): {
-  withTwoslashInlineCache: (config: UserConfig) => UserConfig
-  typesCache: TwoslashTypesCache
-} {
-  const { inject: injectFenceSourceMap, extract: extractFenceSourceMap } = createFenceSourceMap()
+export function createTwoslashWithInlineCache(options: VitePressPluginTwoslashOptions = {}): (config: UserConfig) => UserConfig {
+  return function (config: UserConfig): UserConfig {
+    // provide twoslash cache reader and writer
+    const { typesCache, patcher } = createInlineTypesCache()
 
-  // inject source map to all fences on load
-  const pre: Plugin = {
-    name: 'vitepress-twoslash:fence-source',
-    enforce: 'pre',
-    load(id) {
-      if (id.endsWith('.md')) {
-        const code = readFileSync(id, 'utf-8')
-        return injectFenceSourceMap(code, id)
-      }
-    },
-  }
+    const transformer = transformerTwoslash({ ...options, typesCache })
 
-  // extract and remove source map from fence
-  const transformerFenceSourceMap: ShikiTransformer = {
-    name: '@shikijs/vitepress-twoslash-fence',
-    preprocess(code) {
-      const { code: transformedCode, sourceMap } = extractFenceSourceMap(code)
-      this.meta.sourceMap = sourceMap
-      return transformedCode
-    },
-  }
+    // patch magic files changes after transform
+    const PatchPlugin: Plugin = {
+      name: 'vitepress-twoslash:patch',
+      enforce: 'post',
+      transform(code, id) {
+        if (id.endsWith('.md')) {
+          patcher.patch(id)
+        }
+      },
+    }
 
-  // provide twoslash cache reader and writer
-  const { typesCache, patcher } = createInlineTypesCache()
+    // add source map to markdown fence
+    config = withFenceSourceMap(config)
 
-  // patch magic files changes after transform
-  const post: Plugin = {
-    name: 'vitepress-twoslash:patch',
-    enforce: 'post',
-    transform(code, id) {
-      if (id.endsWith('.md')) {
-        patcher.patch(id)
-      }
-    },
-  }
-
-  function withTwoslashInlineCache(config: UserConfig): UserConfig {
     // config markdown code transformers
     const codeTransformers = (config.markdown ??= {}).codeTransformers ??= []
-    codeTransformers.unshift(transformerFenceSourceMap)
+    codeTransformers.push(transformer)
 
     // config vite plugins
     const plugins = (config.vite ??= {}).plugins ??= []
-    plugins.push(pre, post)
+    plugins.push(PatchPlugin)
 
     return config
-  }
-
-  return {
-    withTwoslashInlineCache,
-    typesCache,
   }
 }
